v7 - Next CRUD Version (#5715)
* Uploaders - Refactor and fixes (#5478)

* add method to get ajax uploaders

* Apply fixes from StyleCI

[ci skip] [skip ci]

* use an abstract class

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* refactor uploaders

* Apply fixes from StyleCI

[ci skip] [skip ci]

* refactor rules

* Apply fixes from StyleCI

[ci skip] [skip ci]

* move ajax to PRO, cleanup

* Apply fixes from StyleCI

[ci skip] [skip ci]

* make attributes available for all subfields

* fix tests

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* upload multiple and upload properly working :pray:

* fixes

* Apply fixes from StyleCI

[ci skip] [skip ci]

* allow the configuration of valueWithoutPath call.

* fix valid upload inside repeatables

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix condition

* cleanup

* fix

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix use case for enabling validation after entry is created

* Apply fixes from StyleCI

[ci skip] [skip ci]

* dont save array keys

* fix ajax validation

* fix validation messages

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fixes ValidUpload

* Apply fixes from StyleCI

[ci skip] [skip ci]

* dont json encode if casted in the model

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix previous file identification in repeatable

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix getting values

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add fake fields support

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip add uploaders tests

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add pro columns

* fix test suite

* fix tests

* ffix tests

* remove unused test views

* add uploaders to test coverage

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add coverage folder to gitignore

* make tests run faster by not reloading db when not necessary

* add coverage to validation tests

* add fake tests to uploaders

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add more tests

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add more upload assets

* fixes

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix single file

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add image column

* fix tests

* Apply fixes from StyleCI

[ci skip] [skip ci]

* remove hardcoded macro names

* Apply fixes from StyleCI

[ci skip] [skip ci]

* remove double loop, fix single file uploader

* Apply fixes from StyleCI

[ci skip] [skip ci]

* use a big increments and unsigned for primary key

* handle pivot file deletion

* Apply fixes from StyleCI

[ci skip] [skip ci]

* register events for relation models

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix typo

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix relationship uploaders

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* update temporary time key

* save objects in the macro

---------


* trigger testing forkflow on PR

* Clean up uploaders (#5725)

* wip

* clean up

* add get uploadedFilesFromRequest to clean up uploaders

* Apply fixes from StyleCI

[ci skip] [skip ci]

* fix type cast

* add summernote uploader

* update summernote

* add summernote

* add more test assets

---------


* allow datatable url configuration (#5713)

* use next branch for basset

* require alpha version

* bump basset version

* Lifecycle hooks for CRUD operations (#5687)

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* tidy up

* Apply fixes from StyleCI

[ci skip] [skip ci]

* panel hooks

* Apply fixes from StyleCI

[ci skip] [skip ci]

* refactoring :broom

* finish implementing hooks

* Apply fixes from StyleCI

[ci skip] [skip ci]

* refactoring hooks

* Apply fixes from StyleCI

[ci skip] [skip ci]

* add hook to delete

* allow empty parameters

* wip

---------


* "Agnostic" filters - decouple filters from datatables (#5714)

* wip

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* wip

* Apply fixes from StyleCI

[ci skip] [skip ci]

* remove filters navbar component

---------


* bump basset version

* Update composer.json

* Apply fixes from StyleCI

[ci skip] [skip ci]

* datatable component (#5688)


* move the list heading to the datatable component (#5808)

* add Datagrid and Datalist components (#5810)


* Form component (no ajax form) (#5814)


* dataform component fixes (#5816)


* Added general chip blade file and chip widget (#5815)

* Fix Form issues, Add *name* argument to forms (#5817)


* Drop support for Laravel 10 and 11 (#5819)

* Fix hooks at setup level (#5818)


* install dev-next version of generators

* require basset beta version (#5821)

* fix scripts on table

* fix the keyboard shortcuts on forms

* ensure proper crud identification

* fix hooks

* fix crud field javacript issue

* pass the form id to form content

* add error page translations

* fix export buttons

* wip

* wip

* wip

* map the ui files

* fix table without export buttons

* always focus on error field

* delete button redirect

* Apply fixes from StyleCI

[ci skip] [skip ci]

* Fix padding in name

* Fix checkbox padding

* Background of select2

* multiple fixes

* fix form error highlight

* wip

* Add toggle to JSON

* allow user to choose locale and timezone

* fixes bulk actions

* fix the filter navbar

* fix line buttons dropdown

* wip

* wip

* wip

* wip

* wip

* DataFormModal component (#5863)


* wip

* wip

* wip

* Save action classes (#5878)

Enable the usage of classes to describe the save actions. This allow more flexibility and easy to extend and/or configure behaviour.

* add cancel button option to modal form (#5881)

* fix problem with checklist_dependency field in Update operation

* fix indentation

* Update src/app/Console/Commands/Install.php

* Update composer.json

---------

+++ src/app/Library/Uploaders/Uploader.php
@@ -2,6 +2,7 @@
 
 namespace Backpack\CRUD\app\Library\Uploaders;
 
+use Backpack\CRUD\app\Library\CrudPanel\CrudPanelFacade as CRUD;
 use Backpack\CRUD\app\Library\Uploaders\Support\Interfaces\UploaderInterface;
 use Backpack\CRUD\app\Library\Uploaders\Support\Traits\HandleFileNaming;
 use Backpack\CRUD\app\Library\Uploaders\Support\Traits\HandleRepeatableUploads;
@@ -21,7 +22,7 @@ abstract class Uploader implements UploaderInterface
 
     private string $path = '';
 
-    private bool $handleMultipleFiles = false;
+    public bool $handleMultipleFiles = false;
 
     private bool $deleteWhenEntryIsDeleted = true;
 
@@ -76,17 +77,19 @@ abstract class Uploader implements UploaderInterface
             return $this->handleRepeatableFiles($entry);
         }
 
+        $values = $this->getUploadedFilesFromRequest();
+
         if ($this->attachedToFakeField) {
             $fakeFieldValue = $entry->{$this->attachedToFakeField};
             $fakeFieldValue = is_string($fakeFieldValue) ? json_decode($fakeFieldValue, true) : (array) $fakeFieldValue;
-            $fakeFieldValue[$this->getAttributeName()] = $this->uploadFiles($entry);
+            $fakeFieldValue[$this->getAttributeName()] = $this->uploadFiles($entry, $values);
 
             $entry->{$this->attachedToFakeField} = isset($entry->getCasts()[$this->attachedToFakeField]) ? $fakeFieldValue : json_encode($fakeFieldValue);
 
             return $entry;
         }
 
-        $entry->{$this->getAttributeName()} = $this->uploadFiles($entry);
+        $entry->{$this->getAttributeName()} = $this->uploadFiles($entry, $values);
 
         return $entry;
     }
@@ -102,16 +105,14 @@ abstract class Uploader implements UploaderInterface
 
     public function deleteUploadedFiles(Model $entry): void
     {
-        if ($this->deleteWhenEntryIsDeleted) {
-            if (! in_array(SoftDeletes::class, class_uses_recursive($entry), true)) {
-                $this->performFileDeletion($entry);
+        if (! in_array(SoftDeletes::class, class_uses_recursive($entry), true)) {
+            $this->performFileDeletion($entry);
 
-                return;
-            }
+            return;
+        }
 
-            if ($entry->isForceDeleting() === true) {
-                $this->performFileDeletion($entry);
-            }
+        if ($entry->isForceDeleting() === true) {
+            $this->performFileDeletion($entry);
         }
     }
 
@@ -153,6 +154,21 @@ abstract class Uploader implements UploaderInterface
         return $this->deleteWhenEntryIsDeleted;
     }
 
+    public function shouldUploadFiles($entryValue): bool
+    {
+        return true;
+    }
+
+    public function shouldKeepPreviousValueUnchanged(Model $entry, $entryValue): bool
+    {
+        return $entry->exists && ($entryValue === null || $entryValue === [null]);
+    }
+
+    public function hasDeletedFiles($entryValue): bool
+    {
+        return $entryValue === false || $entryValue === null || $entryValue === [null];
+    }
+
     public function getIdentifier(): string
     {
         if ($this->handleRepeatableFiles) {
@@ -182,11 +198,30 @@ abstract class Uploader implements UploaderInterface
         if (! $this->attachedToFakeField) {
             return $this->getOriginalValue($entry);
         }
-
         $value = $this->getOriginalValue($entry, $this->attachedToFakeField);
         $value = is_string($value) ? json_decode($value, true) : (array) $value;
 
         return $value[$this->getAttributeName()] ?? null;
+    }
+
+    public function getValueWithoutPath(?string $value = null): ?string
+    {
+        return $value ? Str::after($value, $this->path) : null;
+    }
+
+    public function getUploadedFilesFromRequest()
+    {
+        return CRUD::getRequest()->file($this->getNameForRequest());
+    }
+
+    public function isFake(): bool
+    {
+        return $this->attachedToFakeField !== false;
+    }
+
+    public function getFakeAttribute(): bool|string
+    {
+        return $this->attachedToFakeField;
     }
 
     /*******************************
@@ -206,6 +241,13 @@ abstract class Uploader implements UploaderInterface
         return $this;
     }
 
+    public function fake(bool|string $isFake): self
+    {
+        $this->attachedToFakeField = $isFake;
+
+        return $this;
+    }
+
     /*******************************
      * Default implementation functions
      *******************************/
@@ -217,6 +259,18 @@ abstract class Uploader implements UploaderInterface
     {
         $value = $entry->{$this->getAttributeName()};
 
+        if ($this->attachedToFakeField) {
+            $values = $entry->{$this->attachedToFakeField};
+
+            $values = is_string($values) ? json_decode($values, true) : $values;
+            $attributeValue = $values[$this->getAttributeName()] ?? null;
+            $attributeValue = is_array($attributeValue) ? array_map(fn ($value) => $this->getValueWithoutPath($value), $attributeValue) : $this->getValueWithoutPath($attributeValue);
+            $values[$this->getAttributeName()] = $attributeValue;
+            $entry->{$this->attachedToFakeField} = isset($entry->getCasts()[$this->attachedToFakeField]) ? $values : json_encode($values);
+
+            return $entry;
+        }
+
         if ($this->handleMultipleFiles) {
             if (! isset($entry->getCasts()[$this->getName()]) && is_string($value)) {
                 $entry->{$this->getAttributeName()} = json_decode($value, true);
@@ -225,24 +279,24 @@ abstract class Uploader implements UploaderInterface
             return $entry;
         }
 
-        if ($this->attachedToFakeField) {
-            $values = $entry->{$this->attachedToFakeField};
-            $values = is_string($values) ? json_decode($values, true) : (array) $values;
-
-            $values[$this->getAttributeName()] = isset($values[$this->getAttributeName()]) ? Str::after($values[$this->getAttributeName()], $this->path) : null;
-            $entry->{$this->attachedToFakeField} = json_encode($values);
-
-            return $entry;
-        }
-
-        $entry->{$this->getAttributeName()} = Str::after($value, $this->path);
+        $entry->{$this->getAttributeName()} = $this->getValueWithoutPath($value);
 
         return $entry;
     }
 
-    private function deleteFiles(Model $entry)
+    protected function deleteFiles(Model $entry)
     {
-        $values = $entry->{$this->getAttributeName()};
+        if (! $this->shouldDeleteFiles()) {
+            return;
+        }
+
+        if ($this->attachedToFakeField) {
+            $values = $entry->{$this->attachedToFakeField};
+            $values = is_string($values) ? json_decode($values, true) : $values;
+            $values = $values[$this->getAttributeName()] ?? null;
+        }
+
+        $values ??= $entry->{$this->getAttributeName()};
 
         if ($values === null) {
             return;
@@ -250,7 +304,7 @@ abstract class Uploader implements UploaderInterface
 
         if ($this->handleMultipleFiles) {
             // ensure we have an array of values when field is not casted in model.
-            if (! isset($entry->getCasts()[$this->name]) && is_string($values)) {
+            if (is_string($values)) {
                 $values = json_decode($values, true);
             }
             foreach ($values ?? [] as $value) {
@@ -267,7 +321,7 @@ abstract class Uploader implements UploaderInterface
 
     private function performFileDeletion(Model $entry)
     {
-        if (! $this->handleRepeatableFiles) {
+        if (! $this->handleRepeatableFiles && $this->deleteWhenEntryIsDeleted) {
             $this->deleteFiles($entry);
 
             return;
@@ -277,7 +331,7 @@ abstract class Uploader implements UploaderInterface
     }
 
     /*******************************
-     * Private helper methods
+     * helper methods
      *******************************/
     private function getPathFromConfiguration(array $crudObject, array $configuration): string
     {
